[BE] rename: LogProperty 관련 클래스 패키지 이동  #650
- LogProperty 관련 클래스 패키지 구조 정리 및 이동
