--- backend/core/src/main/java/com/onair/hearit/core/log/dto/LogProperty.java
@@ -1,6 +1,0 @@
-package com.onair.hearit.core.log.dto;
-
-public interface LogProperty {
-
-    LogEvent getlogEvent();
-}
