+++ backend/core/src/main/java/com/onair/hearit/core/log/dto/logproperty/LogProperty.java
@@ -1,0 +1,8 @@
+package com.onair.hearit.core.log.dto.logproperty;
+
+import com.onair.hearit.core.log.dto.LogEvent;
+
+public interface LogProperty {
+
+    LogEvent getlogEvent();
+}
